fix: stabilize sick vacation override
